Move model/mapping utils into packaged code for reuse by the tool shed.
