+++ lib/galaxy/model/unittest_utils/gxy_model_fixtures.py
@@ -2,8 +2,8 @@ import pytest
 
 from galaxy import model
 from galaxy.model.orm.util import add_object_to_object_session
-from . import get_unique_value
-from ...testing_utils import (
+from galaxy.model.unittest_utils.mapping_testing_utils import get_unique_value
+from galaxy.model.unittest_utils.model_testing_utils import (
     dbcleanup_wrapper,
     initialize_model,
 )
