+++ test/unit/data/model/mapping/test_gxy_model_mapping.py
@@ -75,23 +75,21 @@ from sqlalchemy import (
 from galaxy import model
 from galaxy.model.orm.now import now
 from galaxy.model.orm.util import add_object_to_object_session
-from .testing_utils import (
+from galaxy.model.unittest_utils.mapping_testing_utils import (
     AbstractBaseTest,
     collection_consists_of_objects,
     get_unique_value,
     has_index,
     has_unique_constraint,
 )
-from ..testing_utils import (
+from galaxy.model.unittest_utils.model_testing_utils import (
     dbcleanup,
     delete_from_database,
-    get_plugin_full_name,
     get_stored_obj,
     persist,
 )
 
-model_fixtures = get_plugin_full_name("mapping.testing_utils.gxy_model_fixtures")
-pytest_plugins = [model_fixtures]
+pytest_plugins = ["galaxy.model.unittest_utils.gxy_model_fixtures"]
 
 
 class BaseTest(AbstractBaseTest):
